Update after unittests

+++ Exercise_1.py
@@ -10,14 +10,14 @@ book_dict = {
 role_dict = {
     "name":"Gogi",
     "type":"detective",
-    "book": 959,
+    "book": "http://pulse-rest-testing.herokuapp.com/books/6631",
     "level": 100500
 }
 
 def add_new_item(base_url, data_dict):
     """
     Create new item
-    :param base_url: base API urk
+    :param base_url: base API url
     :param data_dict: data of the item (dict)
     :return: new item ID
     """
@@ -31,7 +31,7 @@ def add_new_item(base_url, data_dict):
 
 def add_item_id(item_dict, item_id, key_value="id"):
     """
-    Update item dict wit its id
+    Update item dict with its id
     :return: dict with id value
     """
     item_dict.update({key_value: int(item_id)})
@@ -40,7 +40,7 @@ def add_item_id(item_dict, item_id, key_value="id"):
 def compare_dicts(dict1, dict2):
     unmatched = set(dict1.items()) ^ set(dict2.items())
     if len(unmatched) != 0:
-        raise Exception("The item hasn't been created")
+        raise Exception("Dicts are not equal")
 
 def check_new_item(base_url, item_id, item_dict):
     """
@@ -69,50 +69,49 @@ def check_item_in_list(base_url, item_id, item_dict):
     if len(books_id) != 1:
         raise Exception("The item is not in the list")
 
-def book_update_and_check(base_url, book_id, new_author=book_dict["author"], new_title=book_dict["title"]):
+def book_update_and_check(base_url, book_id, item_dict, new_author=book_dict["author"], new_title=book_dict["title"]):
     """
     Change books data and checks changes in the book structure, and in the all books list.
     """
-    book_dict["author"] = new_author
-    book_dict["title"] = new_title
-    requests.put(base_url + book_id, data=book_dict)
+    item_dict["author"], item_dict["title"] = new_author, new_title
+    requests.put(base_url + book_id, data=item_dict)
     resp_check = requests.get(base_url + book_id)
+    new_item_dict = add_item_id(item_dict, book_id)
     if resp_check.status_code == 200:
         resp = resp_check.json()
-        compare_dicts(resp, book_dict)
+        compare_dicts(resp, new_item_dict)
     else:
         raise Exception("Wrong request")
     list_check = requests.get(base_url)
-    books_id = []
-    for i in list_check.json():
-        if i["id"] == int(book_id):
-            compare_dicts(i, book_dict)
-            books_id.append(i["id"])
-    if len(books_id) != 1:
-        raise Exception("Updated book is not in the list")
+    if list_check.status_code == 200:
+        books_id = []
+        for i in list_check.json():
+            if i["id"] == int(book_id):
+                compare_dicts(i, new_item_dict)
+                books_id.append(i["id"])
+        if len(books_id) != 1:
+            raise Exception("Updated book is not in the list")
 
-def role_update_and_check(base_url, role_id, new_name=role_dict["name"], new_type=role_dict["type"],
-                          new_book=role_dict["book"], new_level=role_dict["level"] ):
+def role_update_and_check(base_url, role_id, item_dict, new_name=role_dict["name"], new_type=role_dict["type"],
+                          new_book=role_dict["book"], new_level=role_dict["level"]):
     """
     Change role data and checks changes in the role structure, and in the all roles list.
     """
-    role_dict["name"] = new_name
-    role_dict["type"] = new_type
-    role_dict["book"] = new_book
-    role_dict["level"] = new_level
-    requests.put(base_url + role_id, data=role_dict)
+    item_dict["name"], item_dict["type"], item_dict["level"], item_dict["book"]= new_name, new_type, new_level, new_book
+    requests.put(base_url + role_id, data=item_dict)
     resp_check = requests.get(base_url + role_id)
+    new_item_dict = add_item_id(item_dict, role_id)
     if resp_check.status_code == 200:
         resp = resp_check.json()
-        compare_dicts(resp, role_dict)
+        compare_dicts(resp, new_item_dict)
     else:
-        raise Exception("The role hasn't been changed")
+        raise Exception("Wrong request")
     list_check = requests.get(base_url)
     if list_check.status_code == 200:
         roles_id = []
         for i in list_check.json():
             if i["id"] == int(role_id):
-                compare_dicts(resp, role_dict)
+                compare_dicts(i, item_dict)
                 roles_id.append(i["id"])
         if len(roles_id) != 1:
             raise Exception("Updated book is not in the list")
@@ -123,7 +122,22 @@ def delete_item_finally(base_url, item_id):
     """
     Delete item
     """
-    requests.delete(base_url + item_id)
+    if item_id is not None:
+       del_resp = requests.delete(base_url + item_id)
+       if del_resp.status_code == 204:
+           list_check = requests.get(base_url)
+           if list_check.status_code == 200:
+               items_id = []
+               for i in list_check.json():
+                   if i["id"] == int(item_id):
+                       items_id.append(i["id"])
+               if len(items_id) != 0:
+                   raise Exception("Something went wrong. Deleted item is still in the list.")
+       else:
+           raise Exception("Wrong request status code. Item hasn't been deleted")
+    else:
+        raise Exception("Item id is None")
+
 
 
 if __name__ == "__main__":
@@ -131,12 +145,12 @@ if __name__ == "__main__":
     book_id = add_new_item(base_url_book, book_dict)
     check_new_item(base_url_book, book_id, book_dict)
     check_item_in_list(base_url_book, book_id, book_dict)
-    book_update_and_check(base_url_book, book_id, "Somerset Moem", "Luna i grosh")
+    book_update_and_check(base_url_book, book_id, book_dict, "Somerset Moem", "Luna i grosh")
     delete_item_finally(base_url_book, book_id)
 
-    # # Manipulations with role
+    # Manipulations with role
     role_id = add_new_item(base_url_role, role_dict)
     check_new_item(base_url_role, role_id, role_dict)
     check_item_in_list(base_url_role,role_id, role_dict)
-    role_update_and_check(base_url_role, role_id, "JuveVR", "student")
+    role_update_and_check(base_url_role, role_id, role_dict, "JuveVR", "student")
     delete_item_finally(base_url_role, role_id)
